feat: Add Card and ContainerCard components with SVG icons for job listings
